Pull out Hierarchy interface
